pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "to_err_is_antenna"
version = "0.1.0"
description = "Visibility Error Simulator for Measurement Sets"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "to_err_is_antenna contributors"}
]
keywords = [
    "radio astronomy",
    "interferometry",
    "measurement set",
    "visibility",
    "calibration",
    "testing",
    "CASA",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Astronomy",
]

dependencies = [
    "numpy>=1.20",
    "pyyaml>=5.4",
    "tqdm>=4.60",
    "python-casacore>=3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
]

[project.scripts]
corruptms = "to_err_is_antenna.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/to_err_is_antenna"
Documentation = "https://github.com/yourusername/to_err_is_antenna#readme"
Repository = "https://github.com/yourusername/to_err_is_antenna.git"
Issues = "https://github.com/yourusername/to_err_is_antenna/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["to_err_is_antenna*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=to_err_is_antenna --cov-report=term-missing"
